requirements.txt
selenium
requests
emoji
pyttsx3
SpeechRecognition
Pillow
pytube